refactor: move lifecycle implementations out of db.rs and into their own modules
